Display traits
Display traits for ParityStorageLocation and ParityLocation, so that
we can get string representations for display to the user.
Serialization is pending, but not critical path since the initial
implementation will be to just report statistics on the size of the
conflict sets via grafana instrumentation, rather than actually
transferring conflict sets.